Extract SmartDate to single class
